requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Migrate
Werkzeug
email-validator
python-dotenv
google-auth requests